+++ revenda-api/src/store/store.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, BadRequestException, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Store } from './store.entity';
@@ -40,7 +40,23 @@ export class StoreService {
     return this.storeRepository.update(id, updateStoreDto);
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    // REGRA DE NEGÓCIO: Não permitir deletar loja que possui vendas associadas
+    const store = await this.storeRepository.findOne({
+      where: { id },
+      relations: ['sales'],
+    });
+
+    if (!store) {
+      throw new NotFoundException(`Loja com ID ${id} não encontrada`);
+    }
+
+    if (store.sales && store.sales.length > 0) {
+      throw new BadRequestException(
+        `Não é possível deletar a loja '${store.name}' pois existem ${store.sales.length} venda(s) associada(s) a ela`
+      );
+    }
+
     return this.storeRepository.delete(id);
   }
 }
